Move redis connection to separate helper
